ROFO-17 refactor: 음식점 hiltModule 분리
